refactor: split models into modules
