+++ backend/app/models/__init__.py
@@ -1,1 +1,31 @@
 """Database models package."""
+
+from .course_stage import CourseStage
+from .goal import Goal
+from .goal_completion import GoalCompletion
+from .goal_group import GoalGroup
+from .habit import Habit
+from .journal_entry import JournalEntry
+from .practice import Practice
+from .practice_session import PracticeSession
+from .prompt_response import PromptResponse
+from .stage_content import StageContent
+from .stage_progress import StageProgress
+from .user import User
+from .user_practice import UserPractice
+
+__all__ = [
+    "CourseStage",
+    "Goal",
+    "GoalCompletion",
+    "GoalGroup",
+    "Habit",
+    "JournalEntry",
+    "Practice",
+    "PracticeSession",
+    "PromptResponse",
+    "StageContent",
+    "StageProgress",
+    "User",
+    "UserPractice",
+]
